requirements.txt
pywebview>=4.4
openpyxl>=3.1
rapidfuzz>=3.9
scikit-learn>=1.5
python-dotenv>=1.0
pydantic>=2.8
PyYAML>=6.0
